refactor(abstraction): move entities validation logic from Contract Generator to Schema Validator (#62)
