+++ packages/abstraction/src/ContractGenerator.ts
@@ -1,7 +1,7 @@
 import { createLogger, Logger } from '@aloxide/logger';
 
 import { readAloxideConfig } from './readAloxideConfig';
-import { validateSchema } from './SchemaValidator';
+import { validateEntity } from './SchemaValidator';
 
 import type { ContractGeneratorConfig } from './ContractGeneratorConfig';
 import type { AloxideConfig } from './AloxideConfig';
@@ -15,15 +15,29 @@ export class ContractGenerator {
       throw new Error('missing configuration');
     }
 
-    if (config.logger) {
-      this.logger = config.logger;
+    const { logger, ...rest } = config;
+    this.config = rest;
+
+    // Define logger for the generator
+    if (logger) {
+      this.logger = logger;
     } else {
       this.logger = createLogger();
     }
     this.logger.debug('-- config.aloxideConfigPath', config.aloxideConfigPath);
     this.logger.debug('-- config.resultPath', config.resultPath);
 
-    this.aloxideConfig = readAloxideConfig(config.aloxideConfigPath);
+    // Check Input config
+    const aloxideConfig = readAloxideConfig(config.aloxideConfigPath);
+    if (!validateEntity(aloxideConfig, this.logger)) {
+      throw new Error('Input entities mismatch!');
+    }
+    this.aloxideConfig = aloxideConfig;
+
+    // Check Output config
+    if (!rest.resultPath) {
+      throw new Error('Missing "resultPath"!');
+    }
 
     if (this.config.adapter) {
       this.config.adapter.logger = this.logger;
@@ -32,41 +46,7 @@ export class ContractGenerator {
     }
   }
 
-  validateEntity() {
-
-    const checkName = val => {
-      return /^[1-5a-zA-Z]+/.test(val)
-    }
-
-    const checkType = val => {
-      const supportedType = ["uint64_t", "number", "string", "array", "bool"];
-      return supportedType.indexOf(val) !== -1
-    }
-
-    const requiredSchema = {
-      entities: [{
-        name: { type: String, required: true, length: { min: 1, max: 12 }, use: { checkName } },
-        fields: [{
-          name: { type: String, required: true, length: { min: 1, max: 12 }, use: { checkName } },
-          type: { type: String, required: true, use: { checkType } },
-        }],
-        key: { type: String, required: true, length: { min: 1, max: 12 }, use: { checkName } }
-      }
-      ]
-    }
-    const schemaErrors = validateSchema(this.aloxideConfig, requiredSchema, this.logger)
-
-    return schemaErrors.length < 1;
-  }
-
   generate() {
-    if (!this.validateEntity()) {
-      throw new Error('input entities mismatch');
-    }
-    if (!this.config.resultPath) {
-      throw new Error('missing resultPath');
-    }
-
     if (this.config.adapter) {
       this.config.adapter.generate(this.config.resultPath);
     }
